+++ crates/nixci/crate.nix
@@ -8,6 +8,7 @@ let
   inherit (flake) inputs;
 in
 {
+  autoWire = true;
   crane = {
     args = {
       nativeBuildInputs = with pkgs; with pkgs.apple_sdk_frameworks; lib.optionals stdenv.isDarwin [
@@ -26,7 +27,12 @@ in
         ) ++ lib.optionals pkgs.stdenv.isLinux [
         pkgs.openssl
       ];
+      # Disable tests due to sandboxing issues; we run them on CI
+      # instead.
+      doCheck = false;
       DEVOUR_FLAKE = inputs.devour-flake;
+      NIX_FLAKE_SCHEMAS_BIN = lib.getExe pkgs.nix-flake-schemas;
+      DEFAULT_FLAKE_SCHEMAS = inputs.flake-schemas;
     };
   };
 }