+++ crates/nixci/src/config/mod.rs
@@ -1,0 +1,5 @@
+//! nixci config defined in `flake.nix`
+pub mod core;
+pub mod ref_;
+pub mod subflake;
+pub mod subflakes;