+++ crates/nixci/src/github/matrix.rs
@@ -1,22 +1,28 @@
-/// Enough types to get branch info from Pull Request URL
+//! Github Actions matrix
 use nix_rs::flake::system::System;
 use serde::{Deserialize, Serialize};
 
-use crate::config::Subflakes;
+use crate::config::subflakes::SubflakesConfig;
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+/// A row in the Github Actions matrix configuration
 pub struct GitHubMatrixRow {
+    /// System to build on
     pub system: System,
+    /// Subflake to build
     pub subflake: String,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+/// Github Actions matrix configuration
 pub struct GitHubMatrix {
+    /// The includes
     pub include: Vec<GitHubMatrixRow>,
 }
 
 impl GitHubMatrix {
-    pub fn from(systems: Vec<System>, subflakes: &Subflakes) -> Self {
+    /// Create a [GitHubMatrix] for the given subflakes and systems
+    pub fn from(systems: Vec<System>, subflakes: &SubflakesConfig) -> Self {
         let include: Vec<GitHubMatrixRow> = systems
             .iter()
             .flat_map(|system| {