+++ crates/nixci/src/github/mod.rs
@@ -1,2 +1,3 @@
+//! GitHub related types and functions.
 pub mod matrix;
 pub mod pull_request;