+++ crates/nixci/src/step/mod.rs
@@ -1,0 +1,4 @@
+//! CI is broken down into various 'steps'.
+pub mod build;
+pub mod core;
+pub mod lockfile;