+++ crates/omnix-cli/crate.nix
@@ -34,12 +34,12 @@ in
         ) ++ lib.optionals pkgs.stdenv.isLinux [
         pkgsStatic.openssl
       ];
-      DEVOUR_FLAKE = inputs.devour-flake;
       OM_INIT_REGISTRY =
         lib.cleanSourceWith {
           name = "flakreate-registry";
           src = flake.inputs.self + /crates/flakreate/registry;
         };
+      DEVOUR_FLAKE = inputs.devour-flake;
       NIX_FLAKE_SCHEMAS_BIN = lib.getExe pkgs.nix-flake-schemas;
       DEFAULT_FLAKE_SCHEMAS = inputs.flake-schemas;
       # Disable tests due to sandboxing issues; we run them on CI