+++ crates/omnix-cli/src/command/completion.rs
@@ -1,7 +1,21 @@
 use crate::Args;
 use clap::CommandFactory;
+use clap::Parser;
 use clap_complete::generate;
 use clap_complete::Shell;
+
+/// Generates shell completion scripts
+#[derive(Parser, Debug)]
+pub struct CompletionCommand {
+    #[arg(value_enum)]
+    shell: clap_complete::Shell,
+}
+
+impl CompletionCommand {
+    pub fn run(&self) -> anyhow::Result<()> {
+        generate_completion(self.shell)
+    }
+}
 
 pub fn generate_completion(shell: Shell) -> anyhow::Result<()> {
     let mut cli = Args::command();