+++ crates/omnix-cli/src/command/mod.rs
@@ -9,25 +9,21 @@ pub mod show;
 
 #[derive(Subcommand, Debug)]
 pub enum Command {
-    Show(show::ShowConfig),
+    Show(show::ShowCommand),
 
-    Init(init::InitConfig),
+    Init(init::InitCommand),
 
-    CI(ci::CIConfig),
+    CI(ci::CICommand),
 
-    Health(health::HealthConfig),
+    Health(health::HealthCommand),
 
-    /// Generates shell completion scripts
-    Completion {
-        #[arg(value_enum)]
-        shell: clap_complete::Shell,
-    },
+    Completion(completion::CompletionCommand),
 }
 
 impl Command {
     pub async fn run(&self, verbosity: Verbosity<InfoLevel>) -> anyhow::Result<()> {
         match self {
-            Command::Completion { shell: _ } => (),
+            Command::Completion(_cmd) => (),
             _ => {
                 let nix_installed = omnix_common::check::nix_installed();
                 if !nix_installed {
@@ -40,11 +36,11 @@ impl Command {
         }
 
         match self {
-            Command::Show(config) => config.run().await,
-            Command::Init(config) => config.run().await,
-            Command::CI(config) => config.run(verbosity).await,
-            Command::Health(config) => config.run().await,
-            Command::Completion { shell } => completion::generate_completion(*shell),
+            Command::Show(cmd) => cmd.run().await,
+            Command::Init(cmd) => cmd.run().await,
+            Command::CI(cmd) => cmd.run(verbosity).await,
+            Command::Health(cmd) => cmd.run().await,
+            Command::Completion(cmd) => cmd.run(),
         }
     }
 }